feature(js): hover menus fire show events and can be refreshed
Hover menus, when displayed, now fire an event `elgg.ui.hoverMenu.show` on the
document. Listeners receive references to the clicked icon and the menu.

The placeholders are now distinct elements which only provide the unique
MAC for the menu and the ajax data to fetch the contents. A menu with the same
MAC either already exists as a child of body, or it's created and loaded. Any
loaded menus can be deleted and will be recreated on demand.